+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # ChangeLog
 
+## 1.0.4 _(October 25, 2014)_
+
+- CLI options
+    - Fixed typo causing `--http-authentication-password` to be ignored.
+- Executables
+    - `arachni_restore` -- Updated to accept timeout options.
+- `Browser`
+    - Fail with `Browser::Error::Spawn` on unsuccessful process spawn.
+- Checks
+    - Active
+        - `csrf` -- Check for `csrf` substring in input names and values.
+    - Passive
+        - `backdoors` -- Added more filenames. [PR #492]
+        - `common_directories` -- Added ISO 3166-1 Alpha-2 countries. [PR #491]
+
 ## 1.0.3 _(October 3, 2014)_
 
 - Added overrides for system write directories in `config/write_paths.yml`.